requirements.txt
streamlit
pandas
psycopg2-binary
pdfplumber
sentence-transformers
scikit-learn
torch
